Split native browser DTO modules
